+++ libclc/clc/include/clc/workitem/clc_get_global_size.h
@@ -9,7 +9,7 @@
 #ifndef __CLC_WORKITEM_CLC_GET_GLOBAL_SIZE_H__
 #define __CLC_WORKITEM_CLC_GET_GLOBAL_SIZE_H__
 
-#include "clc/internal/clc.h"
+#include <clc/internal/clc.h>
 
 _CLC_OVERLOAD _CLC_CONST _CLC_DECL size_t __clc_get_global_size(uint dim);